Extract routes to blueprint
